requirements.txt
Flask
Pillow
google-generativeai
moviepy
grpcio
google-cloud-aiplatform
python-dotenv